fix(about): replace memory description unit
Description: change the memory description unit to GB

Log: 更改内存描述单位GIB为GB
Bug: http://pm.kylin.com/biz/bug-view-28848.html
