+++ plugins/messages-task/about/about.cpp
@@ -21,7 +21,15 @@
 
 #include "about.h"
 #include "ui_about.h"
-#include "memoryentry.h"
+
+#include <KFormat>
+
+#ifdef Q_OS_LINUX
+#include <sys/sysinfo.h>
+#elif defined(Q_OS_FREEBSD)
+#include <sys/types.h>
+#include <sys/sysctl.h>
+#endif
 
 #include <QProcess>
 #include <QFile>
@@ -142,9 +150,8 @@ void About::setupKernelCompenent() {
         cpuType = res["CpuVersion"].toString();
     }
 
-    MemoryEntry memoryInfo;
-    QStringList memory = memoryInfo.totalMemory();
-    memorySize = memorySize + memory.at(0) + "GB"+ "(" + memory.at(1) + "GB" + tr(" available") + ")";
+    QStringList memory = totalMemory();
+    memorySize = memorySize + memory.at(0) + "(" + memory.at(1) + tr(" available") + ")";
 
     ui->cpuContent->setText(cpuType);
     ui->diskContent->setText(diskSize);
@@ -231,6 +238,45 @@ void About::setupSerialComponent() {
 
     connect(ui->activeButton, &QPushButton::clicked, this, &About::runActiveWindow);
     connect(ui->trialButton, &QPushButton::clicked, this, &About::showPdf);
+}
+
+qlonglong About::calculateTotalRam() {
+    qlonglong ret = -1;
+#ifdef Q_OS_LINUX
+    struct sysinfo info;
+    if (sysinfo(&info) == 0)
+        // manpage "sizes are given as multiples of mem_unit bytes"
+        ret = qlonglong(info.totalram) * info.mem_unit;
+#elif defined(Q_OS_FREEBSD)
+    /* Stuff for sysctl */
+    size_t len;
+
+    unsigned long memory;
+    len = sizeof(memory);
+    sysctlbyname("hw.physmem", &memory, &len, NULL, 0);
+
+    ret = memory;
+#endif
+    return ret;
+
+}
+
+QStringList About::totalMemory()
+{
+
+    QStringList res;
+    const qlonglong totalRam = calculateTotalRam();
+
+    if (totalRam > 0) {
+        QString total =  KFormat().formatByteSize(totalRam, 0, KFormat::JEDECBinaryDialect);
+        QString available = KFormat().formatByteSize(totalRam, 1, KFormat::JEDECBinaryDialect);
+        if (atof(total.toLatin1()) < atof(available.toLatin1())) {
+            qSwap(total, available);
+        }
+        res << total << available;
+        return res;
+    }
+    return res;
 }
 
 QStringList About::readFile(QString filepath) {
